Upload new IABot code
This code has a serious bug which makes it unsuitable for deployment.
